requirements.txt
# Core web framework
fastapi>=0.110,<1.0

# ASGI server (with uvloop/httptools/etc)
uvicorn[standard]>=0.24,<1.0

# Database / ORM
SQLAlchemy>=2.0,<3.0
asyncpg>=0.29,<1.0
psycopg2-binary

# Settings & env
pydantic>=2.5,<3.0
pydantic-settings
python-dotenv>=1.0,<2.0
typing-extensions

# Migrations
alembic

# File uploads / async file IO
python-multipart
aiofiles

# Testing & HTTP client
pytest
pytest-asyncio
httpx

# Optional helpers (useful on cloud/platforms)
gunicorn
requests

# Linting / dev (optional)
black
flake8

# AI
langchain_core
langchain_openai
langchain-community
faiss-cpu
openai>=1.0.0
tiktoken
numpy
pandas
supabase
